requirements.txt

langchain_community
langchain-openai
python-dotenv
langchain
langchain_core
streamlit